pyproject.toml
[tool.poetry]
package-mode = false
name = "data-report-testing"
version = "0.1.0"
description = ""
authors = ["xtyuns <xtyuns@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"


[tool.poetry.group.test.dependencies]
pytest-playwright = "^0.4.4"
flake8-pytest-style = "^2.0.0"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
